requirements.txt
pytest~=6.2.2
Appium-Python-Client~=1.1.0
selenium~=3.141.0